requirements.txt
flask
pytesseract
Pillow
PyPDF2
